+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 8.0.0
+#### Notes
+Extends support of the SDK to OneView REST API version 4600 (OneView v8.0.0)
+#### Major changes
+Added support for San Manager resources
+
 # 7.2.0
 #### Notes
 Extends support of the SDK to OneView REST API version 4400 (OneView v7.2.0)